+++ tests/contentstore/test_news_search.py
@@ -68,10 +68,10 @@ def test_news_download(ds_client):
     results = ds_client.contentstore.news.text_search(
         query="",
         max_results=5,
-        search_to_timestamp=datetime.now(timezone.utc) - timedelta(days=100),
+        search_to_timestamp=datetime.now(timezone.utc) - timedelta(days=1),
     )
 
-    content = ds_client.contentstore.news.download(results[-1].id)
+    content = ds_client.contentstore.news.download(results[0].id)
     assert len(content) > 0